+++ ReleaseNotes.md
@@ -152,5 +152,7 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 * New `TextBoxSuggestions.KeepFocusOnSelection` property to control the focus after selecting a suggestion
 * New controls combining a `TextBox` and a `FileSystemDialog`
   * `TextBoxOpenDirectory`
-* New `MaterialWindow.TitleTemplate` property to customize the window title bar
+* Improvements for `MaterialWindow`
+  * New `MaterialWindow.TitleTemplate` property to customize the window title bar
+  * Different looks according to `Window.WindowStyle` property
 * New control `NavigationRail`